requirements.txt
Faker
Flask
python-dotenv
invoke
pytest
pytest-cov
black
pylint